removed another casts from long to int bytelongaccess is now nativetype

+++ src/main/java/net/imglib2/img/unsafe/AbstractUnsafeCursor.java
@@ -134,7 +134,7 @@ public abstract class AbstractUnsafeCursor< T extends NativeLongAccessType< T > 
 	@Override
 	public void jumpFwd( final long steps )
 	{
-		type.incIndex( ( int ) steps );
+		type.incIndex( steps );
 	}
 
 	@Override
