+++ src/main/java/net/imglib2/img/unsafe/UnsafeRandomAccess.java
@@ -62,7 +62,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 		this.img = randomAccess.img;
 		this.type = img.createLinkedType();
 
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; d++ )
 		{
 			position[ d ] = randomAccess.position[ d ];
@@ -117,14 +117,14 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void move( final long distance, final int d )
 	{
-		type.incIndex( img.steps[ d ] * ( int ) distance );
+		type.incIndex( img.steps[ d ] * distance );
 		position[ d ] += distance;
 	}
 
 	@Override
 	public void move( final Localizable localizable )
 	{
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 		{
 			final int distance = localizable.getIntPosition( d );
@@ -137,7 +137,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void move( final int[] distance )
 	{
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 		{
 			position[ d ] += distance[ d ];
@@ -149,7 +149,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void move( final long[] distance )
 	{
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 		{
 			position[ d ] += distance[ d ];
@@ -162,7 +162,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	public void setPosition( final Localizable localizable )
 	{
 		localizable.localize( position );
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 			index += position[ d ] * img.steps[ d ];
 		type.updateIndex( index );
@@ -171,7 +171,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void setPosition( final int[] pos )
 	{
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 		{
 			position[ d ] = pos[ d ];
@@ -183,10 +183,10 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void setPosition( final long[] pos )
 	{
-		int index = 0;
+		long index = 0;
 		for ( int d = 0; d < n; ++d )
 		{
-			final int p = ( int ) pos[ d ];
+			final long p = pos[ d ];
 			position[ d ] = p;
 			index += p * img.steps[ d ];
 		}
@@ -203,8 +203,8 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	@Override
 	public void setPosition( final long pos, final int d )
 	{
-		type.incIndex( ( ( int ) pos - position[ d ] ) * img.steps[ d ] );
-		position[ d ] = ( int ) pos;
+		type.incIndex( ( pos - position[ d ] ) * img.steps[ d ] );
+		position[ d ] = pos;
 	}
 
 	@Override
@@ -259,7 +259,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	 */
 	public void move( final long distance )
 	{
-		type.incIndex( ( int ) distance );
+		type.incIndex( distance );
 		position[ 0 ] += distance;
 	}
 
@@ -289,7 +289,7 @@ public class UnsafeRandomAccess< T extends NativeLongAccessType< T > > extends A
 	 */
 	public void setPositionDim0( final long pos )
 	{
-		type.updateIndex( ( int ) pos );
-		position[ 0 ] = ( int ) pos;
+		type.updateIndex( pos );
+		position[ 0 ] = pos;
 	}
 }
